app/worker/requirements.txt
boto3==1.9.91
mongoengine==0.16.3
python-etcd==0.4.5
Flask==1.0.2